% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpca_con.R
\name{sfpca}
\alias{sfpca}
\title{This function computes the sfpc when the response variable is continuous.}
\usage{
sfpca(xmat, y, theta, lambda, npc_select = 3, timepts = NULL)
}
\arguments{
\item{xmat}{a matrix. Each column represents a sample.}

\item{y}{a vector. The response variable for each sample.}

\item{theta}{between 0 and 1. The weight parameter.}

\item{lambda}{a positive number. The smoothing parameter.}

\item{npc_select}{an integer. The number of FPCs required.}
}
\description{
This function computes the sfpc when the response variable is continuous.
}
\examples{
\dontrun{
load('/Users/joha/Dropbox/Yunlong/Research_ideas/supervisored_FPCA/medfly/application1/medfly.Rdata')
xmat = medfly$eggcount  # 26 days and 50 flies 
y = medfly$lifetime
res = sfpca(xmat,y,0.1,1e3,npc_select=5)
plot(res$beta_fd)
res$fitted 
res$sfpcs
}
}

