% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpca_con.R
\name{sfpca_con}
\alias{sfpca_con}
\title{This function computes the sfpc when the response variable is continuous.}
\usage{
sfpca_con(xmat, y, theta, lambda, npc_select = 3, timepts = NULL,
  xmat_new = NULL)
}
\arguments{
\item{xmat}{a matrix. Each column represents a sample.}

\item{y}{a vector. The response variable for each sample.}

\item{theta}{between 0 and 1. The weight parameter.}

\item{lambda}{a positive number. The smoothing parameter.}

\item{npc_select}{an integer. The number of FPCs required.}

\item{xmat_new}{a matrix. New dataset wish to predict on, each col corresponds to one sample.}
}
\description{
This function computes the sfpc when the response variable is continuous.
}
\examples{
\dontrun{
data(medfly)
xmat = medfly$eggcount  # 26 days and 50 flies 
y = medfly$lifetime
res = sfpca_con(xmat,y,0.1,1e3,npc_select=5,xmat_new = xmat[,1:10])
plot(res$beta_fd,ylab="beta(t)")
res$fitted
res$predicted 
res$sfpcs
}
}

