\name{sfpca-package}
\alias{sfpca-package}
\alias{sfpca}
\docType{package}
\title{
\packageTitle{sfpca}
}
\description{
\packageDescription{sfpca}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{sfpca}
\packageIndices{sfpca}
~~ An overview of how to use the package, ~~
~~ including the most important functions ~~
}
\author{
\packageAuthor{sfpca}

Maintainer: \packageMaintainer{sfpca}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per ~~
~~ line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
